* Mucking around with the TransitionLinks samples, simplifying useWindow to the basics nonhook usage, into a more reusable path extracting getWindow lib method * Adding some more indexing references for lib methods
