+++ src/lib/hooks/index.js
@@ -1,0 +1,8 @@
+import Modal, { modalTypes } from './Modal/Modal'
+import useModal from './useModal'
+
+export {
+  useModal,
+  Modal,
+  modalTypes,
+}
