--- src/lib/hooks/useWindow.js
@@ -1,13 +1,0 @@
-import { useState } from 'react'
-
-const useWindow = () => {
-  const [win, setWin] = useState()
-
-  if (!win && typeof window !== 'undefined') {
-    setWin(window)
-  }
-
-  return win
-}
-
-export default useWindow
