+++ src/lib/index.js
@@ -1,0 +1,8 @@
+import getWindow from './getWindow'
+import { consoleBadge, consoleStyles } from './console'
+
+export {
+  getWindow,
+  consoleBadge,
+  consoleStyles,
+}
