+++ src/lib/static/index.js
@@ -1,0 +1,5 @@
+import useSiteMetadata from './useSiteMetadata'
+
+export {
+  useSiteMetadata
+}
