+++ src/samples/index.js
@@ -2,4 +2,8 @@ import KitchenSink from './KitchenSink/KitchenSink'
 import Modal from './Modal/Modal'
 import TransitionLinks from './TransitionLinks/TransitionLinks'
 
-export { KitchenSink, Modal, TransitionLinks }
+export {
+  KitchenSink,
+  Modal,
+  TransitionLinks,
+}
